completed exercises

+++ P9Operators/Program.cs
@@ -9,7 +9,7 @@ if (int.TryParse(userInput, out var totalSeconds))
     hours -= (24 * days);
     var minutes = (totalSeconds % 3600) / 60;  // (TOT reminder of number of seconds in a day) / number of minutes in an hour
     var seconds = totalSeconds % (minutes * 60); // TOT % (minutes * seconds in a minute)
-    var daysTotal = totalSeconds / (24 * 3600f); // TOT / (hours in a day * seconds in a day) Float for precision
+    var daysTotal = (float) totalSeconds / (24 * 3600); // TOT / (hours in a day * seconds in a day) Float for precision
     
     Console.WriteLine($"""
                        seconds: {seconds}
